Fix make-cjs-components script to work without mjs files being built first
